move suite out of gem entrypoint
